+++ code/CCoA_ij_idea/src/Logic/FootprintSpaceTime/PointClass.java
@@ -1,31 +1,31 @@
 package Logic.FootprintSpaceTime;
 
 public class PointClass implements Point {
-    private int  x;
-    private int  y;
+    private double x;
+    private double y;
 
-    public PointClass(int  x, int  y) {
+    public PointClass(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     @Override
-    public int  getX() {
+    public double getX() {
         return x;
     }
 
     @Override
-    public void setX(int  x) {
+    public void setX(double  x) {
         this.x = x;
     }
 
     @Override
-    public int  getY() {
+    public double getY() {
         return y;
     }
 
     @Override
-    public void setY(int  y) {
+    public void setY(double y) {
         this.y = y;
     }
